MultivariateExpansions

Documentation for MultivariateExpansions.

MultivariateExpansions.MollifiedBasis

MultivariateExpansions.MonicOrthogonalPolynomial

MultivariateExpansions.MultivariateBasis

MultivariateExpansions.OrthogonalPolynomial

MultivariateExpansions.EvalDiff

MultivariateExpansions.EvalDiff

MultivariateExpansions.EvalDiff!

MultivariateExpansions.EvalDiff!

MultivariateExpansions.EvalDiff2

MultivariateExpansions.EvalDiff2

MultivariateExpansions.EvalDiff2!

MultivariateExpansions.EvalDiff2!

MultivariateExpansions.Evaluate

MultivariateExpansions.Evaluate

MultivariateExpansions.Evaluate

MultivariateExpansions.Evaluate!

MultivariateExpansions.Evaluate!

MultivariateExpansions.Evaluate!

MultivariateExpansions.JacobiPolynomial

MultivariateExpansions.MonicJacobiPolynomial

MultivariateExpansions.basesAverage

MultivariateExpansions.basesAverage!

MultivariateExpansions.basisAssembly

MultivariateExpansions.basisAssembly!

MultivariateExpansions.score_loss

MultivariateExpansions.MollifiedBasis — Type

MollifiedBasis(start, basis, moll)

Take a basis and "mollify" it by a mollifier moll. Only affects basis functions of degree start or higher.

One example is Hermite Functions

Example

julia> basis = ProbabilistHermitePolynomial();

julia> moll = SquaredExponential();

julia> mollified_basis = MollifiedBasis(2, basis, moll); # starts mollifying at quadratic term

MultivariateExpansions.MonicOrthogonalPolynomial — Type

Struct representing a monic orthogonal polynomial family via three-term recurrence relation:

\[p_{k+1} = (x - a_k)p_k - b_kp_{k-1}\]

MultivariateExpansions.MultivariateBasis — Type

MultivariateBasis(univariateBases...)

A type to store a multivariate basis as a tuple of univariate bases.

MultivariateExpansions.OrthogonalPolynomial — Type

Struct representing an orthogonal polynomial family via three-term recurrence relation:

\[L_kp_{k+1} = (m_kx - a_k)p_k - b_kp_{k-1}\]

MultivariateExpansions.EvalDiff! — Function

EvalDiff!(eval_space::AbstractMatrix, diff_space::AbstractMatrix, basis::UnivariateBasis, x::AbstractVector)

Evaluate the univariate basis basis and its derivative at x and store the results in eval_space and diff_space, respectively.

Example

julia> eval_space = zeros(3,2);

julia> diff_space = zeros(3,2);

julia> EvalDiff!(eval_space, diff_space, LegendrePolynomial(), [0.5, 0.75])

julia> eval_space
3×2 Matrix{Float64}:
  1.0    1.0
  0.5    0.75
 -0.125  0.34375

julia> diff_space
3×2 Matrix{Float64}:
 0.0  0.0
 1.0  1.0
 1.5  2.25

See also: EvalDiff

MultivariateExpansions.EvalDiff! — Method

EvalDiff!(eval_space, diff_space, basis, pts)

Evaluate all univariate bases and their derivatives associated with a multivariate basis at a set of points.

Arguments

eval_space::NTuple{N, AbstractMatrix}: The output space to store the evaluations, (N,(p_j+1, M))

diff_space::NTuple{N, AbstractMatrix}: The output space to store the derivatives, (N,(p_j+1, M))

basis::MultivariateBasis{N}: The multivariate basis to evaluate

pts::AbstractMatrix: The points to evaluate the basis at, (M,N)

MultivariateExpansions.EvalDiff — Method

EvalDiff(max_degree::Int, basis::UnivariateBasis, x::AbstractVector)

Evaluate the univariate basis basis and its derivative at x and return the result.

Example

julia> eval_space, diff_space = EvalDiff(2, LegendrePolynomial(), [0.5, 0.75]);

julia> eval_space
3×2 Matrix{Float64}:
  1.0    1.0
  0.5    0.75
 -0.125  0.34375

julia> diff_space
3×2 Matrix{Float64}:
 0.0  0.0
 1.0  1.0
 1.5  2.25

See also: EvalDiff!

MultivariateExpansions.EvalDiff — Method

EvalDiff(p, basis, pts)

Evaluate all univariate bases and their derivatives associated with a multivariate basis at a set of points.

Arguments

p::NTuple{N,Int}: The maximum degree of each univariate basis (N)

basis::MultivariateBasis{N}: The multivariate basis to evaluate

pts::AbstractMatrix: The points to evaluate the basis at, (M,N)

See also EvalDiff!.

MultivariateExpansions.EvalDiff2! — Function

EvalDiff2!(eval_space::AbstractMatrix, diff_space::AbstractMatrix, diff2_space::AbstractMatrix, basis::UnivariateBasis, x::AbstractVector)

Evaluate the univariate basis basis and its first two derivatives at x and store the results in eval_space, diff_space and diff2_space, respectively.

Example

julia> eval_space = zeros(3,2);

julia> diff_space = zeros(3,2);

julia> diff2_space = zeros(3,2);

julia> EvalDiff2!(eval_space, diff_space, diff2_space, LegendrePolynomial(), [0.5, 0.75])

julia> eval_space
3×2 Matrix{Float64}:
  1.0    1.0
  0.5    0.75
 -0.125  0.34375

julia> diff_space
3×2 Matrix{Float64}:
 0.0  0.0
 1.0  1.0
 1.5  2.25

julia> diff2_space
3×2 Matrix{Float64}:
 0.0  0.0
 0.0  0.0
 3.0  3.0

See also: EvalDiff2

MultivariateExpansions.EvalDiff2! — Method

EvalDiff2!(eval_space, diff_space, diff2_space, basis, pts)

Evaluate all univariate bases and their first two derivatives associated with a multivariate basis at a set of points.

Arguments

eval_space::NTuple{N, AbstractMatrix}: The output space to store the evaluations, (N,(p_j+1, M))

diff_space::NTuple{N, AbstractMatrix}: The output space to store the first derivatives, (N,(p_j+1, M))

diff2_space::NTuple{N, AbstractMatrix}: The output space to store the second derivatives, (N,(p_j+1, M))

basis::MultivariateBasis{N}: The multivariate basis to evaluate

MultivariateExpansions.EvalDiff2 — Method

EvalDiff2(max_degree::Int, basis::UnivariateBasis, x::AbstractVector)

Evaluate the univariate basis basis and its first two derivatives at x and return the results.

Example

julia> eval_space, diff_space, diff2_space = EvalDiff2(2, LegendrePolynomial(), [0.5, 0.75]);

julia> eval_space
3×2 Matrix{Float64}:
  1.0    1.0
  0.5    0.75
 -0.125  0.34375

julia> diff_space
3×2 Matrix{Float64}:
 0.0  0.0
 1.0  1.0
 1.5  2.25

julia> diff2_space
3×2 Matrix{Float64}:
 0.0  0.0
 0.0  0.0
 3.0  3.0

See also: EvalDiff2!

MultivariateExpansions.EvalDiff2 — Method

EvalDiff2(p, basis, pts)

Evaluate all univariate bases and their first two derivatives associated with a multivariate basis at a set of points.

Arguments

p::NTuple{N,Int}: The maximum degree of each univariate basis (N)

basis::MultivariateBasis{N}: The multivariate basis to evaluate

pts::AbstractMatrix: The points to evaluate the basis at, (M,N)

See also EvalDiff2!.

MultivariateExpansions.Evaluate! — Function

Evaluate!(space::AbstractMatrix, basis::UnivariateBasis, x::AbstractVector)

Evaluate the univariate basis basis at x and store the result in space.

Example

julia> space = zeros(3,2);

julia> Evaluate!(space, LegendrePolynomial(), [0.5, 0.75])

julia> space
3×2 Matrix{Float64}:
  1.0    1.0
  0.5    0.75
 -0.125  0.34375

See also: Evaluate

MultivariateExpansions.Evaluate! — Method

Evaluate!(out, fmset, univariateEvals)

Evaluate the basis of a multivariate expansion given the evaluations of the univariate bases

Arguments

out::AbstractMatrix: The output matrix to store the expansion evaluations, (N,M)

fmset::FixedMultiIndexSet{d}: The fixed multi-index set defining the expansion space (N,d)

univariateEvals::NTuple{d, AbstractMatrix}: The univariate evaluations at each marginal point (d,(pj, M)), where pj is the maximum degree of the j-th univariate basis

See also Evaluate.

MultivariateExpansions.Evaluate! — Method

Evaluate!(eval_space, basis, pts)

Evaluate all univariate bases associated with a multivariate basis at a set of points.

Arguments

eval_space::NTuple{N, AbstractMatrix}: The output space to store the evaluations, (N,(p_j+1, M))

basis::MultivariateBasis{N}: The multivariate basis to evaluate

pts::AbstractMatrix: The points to evaluate the basis at, (M,N)

MultivariateExpansions.Evaluate — Method

Evaluate(max_degree::Int, basis::UnivariateBasis, x::AbstractVector)

Evaluate the univariate basis basis at x and return the result.

Example

julia> Evaluate(2, LegendrePolynomial(), [0.5, 0.75])
3×2 Matrix{Float64}:
  1.0    1.0
  0.5    0.75
 -0.125  0.34375

See also: Evaluate!

MultivariateExpansions.Evaluate — Method

Evaluate(p, basis, pts)

Evaluate all univariate bases associated with a multivariate basis at a set of points.

Arguments

p::NTuple{N,Int}: The maximum degree of each univariate basis (N)

basis::MultivariateBasis{N}: The multivariate basis to evaluate

pts::AbstractMatrix: The points to evaluate the basis at, (M,N)

See also Evaluate!.

MultivariateExpansions.Evaluate — Method

Evaluate(fmset, univariateEvals)

Out-of-place evaluation of multivariate expansion

See Evaluate! for details.

MultivariateExpansions.JacobiPolynomial — Method

JacobiPolynomial(α,β)

Jacobi polynomials P^(α,β)_k, orthogonal on [-1,1] with weight (1-x)^α(1+x)^β

MultivariateExpansions.MonicJacobiPolynomial — Method

MonicJacobiPolynomial(α,β)

Monic Jacobi polynomials P^(α,β)_k, orthogonal on [-1,1] with weight (1-x)^α(1+x)^β

MultivariateExpansions.basesAverage! — Method

basesAverage!(out, fmset, univariateEvals)

Evaluate the average of a set of basis functions over a set of points given the univariate basis evaluations

Arguments

out::AbstractVector: The output vector to store the basis averages, (N,)

fmset::FixedMultiIndexSet{d}: The fixed multi-index set defining the function space (N,d)

univariateEvals::NTuple{d, AbstractMatrix}: The univariate evaluations at each marginal point (d,(pj, M)), where pj is the maximum degree of the j-th univariate basis

See also basesAverage.

MultivariateExpansions.basesAverage — Method

basesAverage(fmset, univariateEvals)

Out-of-place evaluation of the average of a set of basis functions

See basesAverage! for details.

MultivariateExpansions.basisAssembly! — Method

basisAssembly!(out, fmset, coeffs, univariateEvals)

Evaluate a multivariate expansion on a set of points given the coefficients and univariate bases evaluated at each point.

Arguments

out::AbstractVector: The output vector to store the expansion evaluations, (M,)

fmset::FixedMultiIndexSet{d}: The fixed multi-index set defining the expansion space (N,d)

coeffs::AbstractVector: The coefficients of the multivariate expansion (N,)

univariateEvals::NTuple{d, AbstractMatrix}: The univariate evaluations at each marginal point (d,(pj, M)), where pj is the maximum degree of the j-th univariate basis

MultivariateExpansions.basisAssembly — Method

basisAssembly(fmset, coeffs, univariateEvals)

Out-of-place assembly of multivariate basis

See basisAssembly! for details.

MultivariateExpansions.score_loss — Method

score_loss(fmset, coeffs, univariateEval, univariateDiff, univariateDiff2)