MultivariateExpansions

Documentation for MultivariateExpansions.

MultivariateExpansions.polynomialAssembly

MultivariateExpansions.polynomialAssembly!

MultivariateExpansions.polynomialsAverage

MultivariateExpansions.polynomialsAverage!

MultivariateExpansions.polynomialsEval

MultivariateExpansions.polynomialsEval!

MultivariateExpansions.score_loss

MultivariateExpansions.polynomialAssembly! — Method

polynomialAssembly!(out, fmset, coeffs, univariateEvals)

Evaluate a polynomial expansion at a set of points given the coefficients and univariate evaluations at each marginal point.

Arguments

out::AbstractVector: The output vector to store the polynomial evaluations, (M,)

fmset::FixedMultiIndexSet{d}: The fixed multi-index set defining the polynomial space (N,d)

coeffs::AbstractVector: The coefficients of the polynomial expansion (N,)

univariateEvals::NTuple{d, AbstractMatrix}: The univariate evaluations at each marginal point (d,(pj, M)), where pj is the maximum degree of the j-th univariate polynomial

MultivariateExpansions.polynomialAssembly — Method

polynomialAssembly(fmset, coeffs, univariateEvals)

Out-of-place assembly of multivariate polynomial

See polynomialAssembly! for details.

MultivariateExpansions.polynomialsAverage! — Method

polynomialsAverage!(out, fmset, univariateEvals)

Evaluate the average of a set of polynomials over a set of points given the univariate polynomial evaluations

Arguments

out::AbstractVector: The output vector to store the polynomial evaluations, (N,)

fmset::FixedMultiIndexSet{d}: The fixed multi-index set defining the polynomial space (N,d)

univariateEvals::NTuple{d, AbstractMatrix}: The univariate evaluations at each marginal point (d,(pj, M)), where pj is the maximum degree of the j-th univariate polynomial

See also polynomialsAverage.

MultivariateExpansions.polynomialsAverage — Method

polynomialsAverage(fmset, univariateEvals)

Out-of-place evaluation of the average of a set of polynomials

See polynomialsAverage! for details.

MultivariateExpansions.polynomialsEval! — Method

polynomialsEval!(out, fmset, univariateEvals)

Evaluate a set of multivariate polynomials given the evaluations of the univariate polynomials

Arguments

out::AbstractMatrix: The output matrix to store the polynomial evaluations, (N,M)

fmset::FixedMultiIndexSet{d}: The fixed multi-index set defining the polynomial space (N,d)

univariateEvals::NTuple{d, AbstractMatrix}: The univariate evaluations at each marginal point (d,(pj, M)), where pj is the maximum degree of the j-th univariate polynomial

See also polynomialsEval.

MultivariateExpansions.polynomialsEval — Method

polynomialsEval(fmset, univariateEvals)

Out-of-place evaluation of multivariate polynomials

See polynomialsEval! for details.

MultivariateExpansions.score_loss — Method

score_loss(fmset, coeffs, univariateEval, univariateDiff, univariateDiff2)