MultivariateExpansions

Documentation for MultivariateExpansions.

MultivariateExpansions.MultivariateBasis

MultivariateExpansions.basesAverage

MultivariateExpansions.basesAverage!

MultivariateExpansions.basisAssembly

MultivariateExpansions.basisAssembly!

MultivariateExpansions.score_loss

UnivariateApprox.EvalDiff

UnivariateApprox.EvalDiff!

UnivariateApprox.EvalDiff2

UnivariateApprox.EvalDiff2!

UnivariateApprox.Evaluate

UnivariateApprox.Evaluate

UnivariateApprox.Evaluate!

UnivariateApprox.Evaluate!

MultivariateExpansions.MultivariateBasis — Type

MultivariateBasis(univariateBases...)

A type to store a multivariate basis as a tuple of univariate bases.

MultivariateExpansions.basesAverage! — Method

basesAverage!(out, fmset, univariateEvals)

Evaluate the average of a set of basis functions over a set of points given the univariate basis evaluations

Arguments

out::AbstractVector: The output vector to store the basis averages, (N,)

fmset::FixedMultiIndexSet{d}: The fixed multi-index set defining the function space (N,d)

univariateEvals::NTuple{d, AbstractMatrix}: The univariate evaluations at each marginal point (d,(pj, M)), where pj is the maximum degree of the j-th univariate basis

See also basesAverage.

MultivariateExpansions.basesAverage — Method

basesAverage(fmset, univariateEvals)

Out-of-place evaluation of the average of a set of basis functions

See basesAverage! for details.

MultivariateExpansions.basisAssembly! — Method

basisAssembly!(out, fmset, coeffs, univariateEvals)

Evaluate a multivariate expansion on a set of points given the coefficients and univariate bases evaluated at each point.

Arguments

out::AbstractVector: The output vector to store the expansion evaluations, (M,)

fmset::FixedMultiIndexSet{d}: The fixed multi-index set defining the expansion space (N,d)

coeffs::AbstractVector: The coefficients of the multivariate expansion (N,)

univariateEvals::NTuple{d, AbstractMatrix}: The univariate evaluations at each marginal point (d,(pj, M)), where pj is the maximum degree of the j-th univariate basis

MultivariateExpansions.basisAssembly — Method

basisAssembly(fmset, coeffs, univariateEvals)

Out-of-place assembly of multivariate basis

See basisAssembly! for details.

MultivariateExpansions.score_loss — Method

score_loss(fmset, coeffs, univariateEval, univariateDiff, univariateDiff2)

UnivariateApprox.EvalDiff! — Method

EvalDiff!(eval_space, diff_space, basis, pts)

Evaluate all univariate bases and their derivatives associated with a multivariate basis at a set of points.

Arguments

eval_space::NTuple{N, AbstractMatrix}: The output space to store the evaluations, (N,(p_j+1, M))

diff_space::NTuple{N, AbstractMatrix}: The output space to store the derivatives, (N,(p_j+1, M))

basis::MultivariateBasis{N}: The multivariate basis to evaluate

pts::AbstractMatrix: The points to evaluate the basis at, (M,N)

UnivariateApprox.EvalDiff — Method

EvalDiff(p, basis, pts)

Evaluate all univariate bases and their derivatives associated with a multivariate basis at a set of points.

Arguments

p::NTuple{N,Int}: The maximum degree of each univariate basis (N)

basis::MultivariateBasis{N}: The multivariate basis to evaluate

pts::AbstractMatrix: The points to evaluate the basis at, (M,N)

See also EvalDiff!.

UnivariateApprox.EvalDiff2! — Method

EvalDiff2!(eval_space, diff_space, diff2_space, basis, pts)

Evaluate all univariate bases and their first two derivatives associated with a multivariate basis at a set of points.

Arguments

eval_space::NTuple{N, AbstractMatrix}: The output space to store the evaluations, (N,(p_j+1, M))

diff_space::NTuple{N, AbstractMatrix}: The output space to store the first derivatives, (N,(p_j+1, M))

diff2_space::NTuple{N, AbstractMatrix}: The output space to store the second derivatives, (N,(p_j+1, M))

basis::MultivariateBasis{N}: The multivariate basis to evaluate

UnivariateApprox.EvalDiff2 — Method

EvalDiff2(p, basis, pts)

Evaluate all univariate bases and their first two derivatives associated with a multivariate basis at a set of points.

Arguments

p::NTuple{N,Int}: The maximum degree of each univariate basis (N)

basis::MultivariateBasis{N}: The multivariate basis to evaluate

pts::AbstractMatrix: The points to evaluate the basis at, (M,N)

See also EvalDiff2!.

UnivariateApprox.Evaluate! — Method

Evaluate!(out, fmset, univariateEvals)

Evaluate the basis of a multivariate expansion given the evaluations of the univariate bases

Arguments

out::AbstractMatrix: The output matrix to store the expansion evaluations, (N,M)

fmset::FixedMultiIndexSet{d}: The fixed multi-index set defining the expansion space (N,d)

univariateEvals::NTuple{d, AbstractMatrix}: The univariate evaluations at each marginal point (d,(pj, M)), where pj is the maximum degree of the j-th univariate basis

See also Evaluate.

UnivariateApprox.Evaluate! — Method

Evaluate!(eval_space, basis, pts)

Evaluate all univariate bases associated with a multivariate basis at a set of points.

Arguments

eval_space::NTuple{N, AbstractMatrix}: The output space to store the evaluations, (N,(p_j+1, M))

basis::MultivariateBasis{N}: The multivariate basis to evaluate

pts::AbstractMatrix: The points to evaluate the basis at, (M,N)

UnivariateApprox.Evaluate — Method

Evaluate(p, basis, pts)

Evaluate all univariate bases associated with a multivariate basis at a set of points.

Arguments

p::NTuple{N,Int}: The maximum degree of each univariate basis (N)

basis::MultivariateBasis{N}: The multivariate basis to evaluate

pts::AbstractMatrix: The points to evaluate the basis at, (M,N)

See also Evaluate!.

UnivariateApprox.Evaluate — Method

Evaluate(fmset, univariateEvals)

Out-of-place evaluation of multivariate expansion

See Evaluate! for details.